Revert tree path tracking and add alignment tests
I started to track the tree structure during graph parsing but did not
continue that method, but accidentally left the code in in a previous
commit. It is now removed.

I updated the alignment code to explicitly disallow alignment on
expanded nodes (e.g., `(a / abc)~e.1`).

Finally, I changed the Graph class so that the set of variables is
calculated on Graph construction instead of each time Graph.variables()
is called. This should be fine until it's possible to modify Graphs
in-place (in which case the variables just need to be regenerated).